+++ routes/index.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
-//const service = require('../service/mysqlService')
-const service = require('../service/mongoService')
+const service = require('../service/mysqlService')
+//const service = require('../service/mongoService')
 
 router.get('/', async function (req, res, next) {
   const categories = await service.getCategories();
@@ -9,3 +9,4 @@ router.get('/', async function (req, res, next) {
 })
 
 module.exports = router;
+
